requirements.txt
praw
openrouter
python-dotenv
